pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "agentverify"
version = "0.1.0"
description = "Deterministic testing for AI agents. Assert tool calls, not vibes."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "simukappu"},
]
keywords = [
    "pytest",
    "testing",
    "ai",
    "agent",
    "llm",
    "tool-calls",
    "cassette",
    "deterministic",
    "ci",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Framework :: Pytest",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Testing",
]
dependencies = [
    "PyYAML",
    "pytest",
]

[project.urls]
Homepage = "https://github.com/simukappu/agentverify"
Repository = "https://github.com/simukappu/agentverify"
Issues = "https://github.com/simukappu/agentverify/issues"
Changelog = "https://github.com/simukappu/agentverify/blob/main/CHANGELOG.md"

[project.optional-dependencies]
openai = ["openai"]
anthropic = ["anthropic"]
gemini = ["google-genai"]
bedrock = ["boto3"]
litellm = ["litellm"]
all = [
    "openai",
    "anthropic",
    "google-generativeai",
    "boto3",
    "litellm",
]
dev = [
    "hypothesis",
    "pytest-cov",
]

[project.entry-points.pytest11]
agentverify = "agentverify.plugin"

[tool.setuptools.packages.find]
include = ["agentverify*"]

[tool.setuptools.package-data]
agentverify = ["py.typed"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = ["-p", "no:agentverify"]
markers = [
    "agentverify: mark test as an agentverify agent test",
]

[tool.coverage.run]
source = ["agentverify"]
branch = true

[tool.coverage.report]
fail_under = 100
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.",
]
